Improvements for mobile
