+++ src/App.css
@@ -72,9 +72,9 @@
 .navigation-links ul {
   list-style-type: none;
   display: flex;
+  flex-direction:column;
   justify-content: right;
-  gap:20px;
-  margin-right:50px;
+  margin-right:30px;
 }
 
 .navigation-links ul li {
@@ -94,8 +94,8 @@
   display: flex;
   flex-flow:row wrap;
   flex-direction: row;
-  justify-content: space-around;
-  gap:30px;
+  justify-content: left;
+  gap:12px;
 }
 
 .clock-container {
@@ -221,7 +221,7 @@
   margin-top:20px;
 }
 
-@media only screen and (max-width: 600px) {
+@media only screen and (max-width: 700px) {
 
   .page {
     padding:0;
@@ -243,7 +243,30 @@
     border-radius:0%;
     text-align:left;
     padding-top:10px;
+    padding-left:10px;
     font-size:0.8em;
   }
 
+  .clock-manage-button-container {
+    justify-content:left;
+  }
+
+  .time-col-container {
+    padding-top:0px;
+  }
+
+  .time-stamp-display {
+    font-size:2.2em;
+  }
+
+  .navigation-links ul {
+    margin-right:2px;
+  }
+}
+
+@media only screen and (max-width: 400px) {
+  .browser-time {
+    font-size:1.2em;
+    font-weight:bold;
+  }
 }
