+++ src/App.tsx
@@ -5,8 +5,6 @@ import Routes from './Routes';
 
 import './App.css';
 
-import * as $ from 'jquery';
-
 function App() {
   return (
     <SettingsProvider>
